Allow non HPX projects to be used with pycicle
Project configuration files have been moved into a subdirectory
corresponding to the project name so that multiple
projects can be supported.

A new project config file is required to set a few variables
such as the project name/repo/organisation on github. These
were previously hardcoded, but can now be set on a per project basis.

If two projects have PR numbers that are the same, then problems
will arise since we only use the PR number and not the project name
as well in the build/src subdirs on the build machine - this will
need to be addressed in a subsequent patch.