pycicle now generates cmake strings for options
Options and dependent options specified in the project file
are correctly parsed, then randomly selected from the possible choices
and then passed onto cmake for the build step.

Currently the random selection of options only happens at startup
and it needs to be done on every build trigger for each PR
independently.